requirements.txt
flask
flask-cors
langchain
langchain-community
langchain-huggingface
langchain-groq
langchain-text-splitters
faiss-cpu
pypdf
python-docx
python-dotenv
werkzeug
gunicorn
